chore: CR merged auth error -> auth os
