+++ src/auth/index.ts
@@ -1,4 +1,3 @@
-export * from './authError'
 export * from './authOS'
 export * from './backgroundTimeout'
 export * from './ChangePinScreen'
